Refactor two web-based scrapers
