+++ source/docq/manage_spaces.py
@@ -73,16 +73,7 @@ def list_documents(space: SpaceKey) -> list[tuple[str, int, int]]:
     space_data_source = SpaceDataSources[ds_type].value
 
     try:
-        if isinstance(space_data_source, (SpaceDataSourceWebBased)):
-            documents_list = space_data_source.get_document_list(space, ds_configs, get_index_dir(space))
-        elif isinstance(space_data_source, (SpaceDataSourceFileBased)):
-            documents_list = space_data_source.get_document_list(space, ds_configs)
-        else:
-            log.warning(
-                "This category of SpaceDataSource class doesn't support listing documents. data source type is %s",
-                ds_type,
-            )
-            documents_list = []
+        documents_list = space_data_source.get_document_list(space, ds_configs)
 
     except Exception as e:
         log.error(
